Ignore missing records in Active Storage migration
There could be inconsistencies in the database and an attachment might
have a `record_id` pointing to a record which no longer exist. We were
getting an exception in this case.

+++ lib/tasks/active_storage.rake
@@ -69,7 +69,7 @@ namespace :active_storage do
     ActiveStorage::Attachment.find_each do |attachment|
       dest = ActiveStorage::Blob.service.path_for(attachment.blob.key)
 
-      next if File.exist?(dest)
+      next if File.exist?(dest) || !attachment.record
 
       name = attachment.name.delete_prefix("storage_")
       source = attachment.record.send(name).path
